test: move serde_custom_coder() test into tests/
