+++ sdks/rust/src/coders/mod.rs
@@ -113,16 +113,17 @@ pub enum Context {
     NeedsDelimiters,
 }
 
+#[macro_export]
 macro_rules! register_coders {
     ($($coder:ident),*) => {
-        fn encode_from_urn(urn: &str, elem: &dyn crate::elem_types::ElemType, writer: &mut dyn std::io::Write, context: &crate::coders::Context) -> Result<usize, std::io::Error> {
+        fn encode_from_urn(urn: &str, elem: &dyn $crate::elem_types::ElemType, writer: &mut dyn std::io::Write, context: &$crate::coders::Context) -> Result<usize, std::io::Error> {
             match urn {
                 $($coder::URN => $coder.encode(elem, writer, context),)*
                 _ => panic!("unknown urn"),
             }
         }
 
-        fn decode_from_urn(urn: &str, reader: &mut dyn std::io::Read, context: &crate::coders::Context) -> Result<Box<dyn crate::elem_types::ElemType>, std::io::Error> {
+        fn decode_from_urn(urn: &str, reader: &mut dyn std::io::Read, context: &$crate::coders::Context) -> Result<Box<dyn $crate::elem_types::ElemType>, std::io::Error> {
             match urn {
                 $($coder::URN => $coder.decode(reader, context),)*
                 _ => panic!("unknown urn"),
